% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_data.R
\name{odm_data}
\alias{odm_data}
\title{Create ODM formated dataframe}
\usage{
odm_data(
  date_time,
  data_value,
  utc_offset = -5,
  site_id,
  variable_id,
  qualifier_id = NA,
  method_id,
  source_id = 1,
  level_id = 0,
  channel = ODM
)
}
\arguments{
\item{date_time}{local date time for each record}

\item{data_value}{matching data value for each date time record}

\item{utc_offset}{difference in hours from Coordinated Universal Time (UTC)}

\item{site_id}{corresponds to an existing Sites record in the Sites table}

\item{variable_id}{corresponds to an existing record in the Variables table}

\item{qualifier_id}{corresponds to an existing record in the Qualifiers table}

\item{method_id}{corresponds to an existing record in the Methods table}

\item{source_id}{corresponds to an existing record in the Sources table}

\item{level_id}{corresponds to a record in the QualityControlLevels table}

\item{channel}{connection handle as returned by odbcConnect}
}
\description{
The function odm_data can be used to help in formating new data. odm_data
takes a vector of dates and data and outputs a dataframe formated for use
with the functions in this package. site_id, variable_id, and method_id need
to already exist in their respective tables in the ODM.
}
\details{
The output is a properly formated dataframe suitable for further
use in workflows involving this package.
}
\examples{

\dontrun{
# dates
date = seq(as.Date("2010/1/1"), as.Date("2011/1/1"), "days")
# dummy data values
value = 1:366

# creation of dataframe that meets the requirements of the ODM.
tmp = odm_data(date_time = date, data_value = value, site_id = 1,
 variable_id = 1, method_id = 9)
}

}
