% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_delete.R
\name{odm_delete}
\alias{odm_delete}
\title{Delete records from ODM database}
\usage{
odm_delete(
  site_id,
  variable_id,
  method_id,
  level_id,
  source_id = 1,
  start_date = "1970-01-1 00:00:00",
  end_date = Sys.Date(),
  channel = ODM
)
}
\arguments{
\item{site_id}{index value for the location at which the observation was made}

\item{variable_id}{index value for the variable that the data represents}

\item{method_id}{index value for the method used to collect the observation}

\item{level_id}{the level of quality control processing}

\item{source_id}{index value for the source of the observation}

\item{start_date}{access data from this date forward}

\item{end_date}{access data up to this date}

\item{channel}{connection handle as returned by odbcConnect}
}
\description{
The function odm_delete can be used to delete data from ODM database.
}
\details{
A standard SQL query is issued to the ODM database and the values are
removed. The data deleted can be limited by a start
and/or end date defined by the user.
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "update", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Delete data by site, variable, method and QC level.
odm_delete(site_id = 1, variable_id = 1, method_id = 9,
 level_id = 0, start_date = "2013-06-01", end_date = "2013-07-01",
 channel = ODM)
 }

}
