% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_read.R
\name{odm_read}
\alias{odm_read}
\title{Query an ODM database}
\usage{
odm_read(
  utc_offset = -5,
  site_id = NULL,
  variable_id = NULL,
  qualifier_id = NULL,
  method_id = NULL,
  source_id = NULL,
  level_id = NULL,
  aggregate_by = "day",
  FUN = "mean",
  start_date = NULL,
  end_date = NULL,
  n = 1e+06,
  channel = ODM
)
}
\arguments{
\item{utc_offset}{of the observation}

\item{site_id}{index value for the location at which the observation was made}

\item{variable_id}{index value for the variable that the data represents}

\item{qualifier_id}{of the observations you want to extract}

\item{method_id}{index value for the method used to collect the observation}

\item{source_id}{the source of the data}

\item{level_id}{the level of quality control processing}

\item{aggregate_by}{aggregation level to include in SQL query. Either hour,
day, month, year or none}

\item{FUN}{function to aggregate by. Either min, max, mean, or sum}

\item{start_date}{access data from this date forward}

\item{end_date}{access data up to this date}

\item{n}{maximum number of observations to return}

\item{channel}{connection handle as returned by odbcConnect}
}
\value{
A data frame
}
\description{
The function odm_read can be used to get data from ODM database. ODMselect
returns a data frame containing the required columns for working with ODM
data.
}
\details{
A standard SQL query is issued to the ODM database and the values are
returned. The data returned can be limited by a start
and/or end date defined by the user. Smaller datasets can speed up the
workflow especially when visualising data.
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "update", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Extract data by site, variable, method and QC level. Clear and consistent
# method to extract data.
Data <- odm_read(site_id = 1, variable_id = 1, method_id = 9,
 level_id = 0, start_date = "2013-06-01",
 end_date = "2013-07-01", channel = ODM)
}

}
