% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_read_tbl.R
\name{odm_read_tbl}
\alias{odm_read_tbl}
\title{ODM read function}
\usage{
odm_read_tbl(odm_tbl = "Seriescatalog", ID = NULL, channel = ODM)
}
\arguments{
\item{odm_tbl}{name of the table in the ODM database}

\item{ID}{id of entry in the table you are referencing}

\item{channel}{connection handle as returned by odbcConnect
(default value = ODM)}
}
\value{
A data frame
}
\description{
This function can retrieve meta data from an ODM database by table
name and ID
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "update", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Get a list of sites in the database
odm_read_tbl(odm_tbl = "Sites")

# Get a list of data series in the database
odm_read_tbl(odm_tbl = "Seriescatalog")
}

}
