% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_tools.R
\name{odm_tools}
\alias{odm_tools}
\title{Launch shiny app to work interactively with ODM data}
\usage{
odm_tools()
}
\description{
The function odm_tools can be used to launch a web based tool to plot and
correct data in an ODM database.
}
\examples{

\dontrun{
odm_tools()
}

}
