% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMcreate.R
\name{ODMcreate}
\alias{ODMcreate}
\title{Create ODM formated dataframe}
\usage{
ODMcreate(LocalDateTime, DataValue, UTCOffset = -5, SiteID, VariableID,
  QualifierID = NA, MethodID, SourceID = 1, QCLevelID = 0)
}
\arguments{
\item{LocalDateTime}{local date time for each record}

\item{DataValue}{matching data value for each date time record}

\item{UTCOffset}{difference in hours from Coordinated Universal Time (UTC)}

\item{SiteID}{corresponds to an existing Sites record in the Sites table}

\item{VariableID}{corresponds to an existing record in the Variables table}

\item{QualifierID}{corresponds to an existing record in the Qualifiers table}

\item{MethodID}{corresponds to an existing record in the Methods table}

\item{SourceID}{corresponds to an existing record in the Sources table}

\item{QCLevelID}{corresponds to a record in the QualityControlLevels table}
}
\description{
The function ODMcreate can be used to help in formating new data. ODMcreate
takes a vector of dates and data and outputs a dataframe formated for use
with the functions in this package. Site ID, VariableID, and MethodID need to
already exist in their respective tables in the ODM.
}
\details{
The vectors are given with the arguments for SiteID, VariableID and MethodID.
The output is a properly formated dataframe suitable for further use in
workflows involving this package.
}
\examples{
# dates
date = seq(as.Date("2010/1/1"), as.Date("2011/1/1"), "days")
# dummy data values
value = 1:366

# creation of formated xts object
tmp = ODMcreate(LocalDateTime = date, DataValue = value, SiteID = 1,
 VariableID = 1, MethodID = 9)

}

