% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMdeleteData.R
\name{ODMdeleteData}
\alias{ODMdeleteData}
\title{Delete records from ODM database}
\usage{
ODMdeleteData(SeriesID = NULL, SiteID = "SiteID",
  VariableID = "VariableID", MethodID = "MethodID",
  QCLevelID = "QualityControlLevelID", startDate = "1970-01-1 00:00:00",
  endDate = Sys.Date(), channel = ODM)
}
\arguments{
\item{SeriesID}{index value for a specific record within the series catalog}

\item{SiteID}{index value for the location at which the observation was made}

\item{VariableID}{index value for the variable that the data represents}

\item{MethodID}{index value for the method used to collect the observation}

\item{QCLevelID}{the level of quality control processing}

\item{startDate}{access data from this date forward}

\item{endDate}{access data up to this date}

\item{channel}{connection handle as returned by odbcConnect}
}
\description{
The function ODMdeleteData can be used to delete data from ODM database.
}
\details{
A standard SQL query is issued to the ODM database and the values are
removed. The data deleted can be limited by a start
and/or end date defined by the user.

Data series from the catalog may be specified using the series catalog
ID. Since the series catalog is not stable through time you may also delete
values by specifying SiteID, VariableID and MethodID.
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "update", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Delete data by SeriesID
ODMdeleteData(SeriesID = 10, startDate = "2013-06-01",
 endDate = "2013-07-01", channel = ODM)

# Delete data by site, variable, method and QC level.
ODMdeleteData(SiteID = 1, VariableID = 1, MethodID = 9,
 QCLevelID = 0, startDate = "2013-06-01", endDate = "2013-07-01",
 channel = ODM)
 }

}
