% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMget.R
\name{ODMgetCatalog}
\alias{ODMgetCatalog}
\title{ODM utility functions}
\usage{
ODMgetCatalog(channel = ODM, ID = NA)
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect (default value = ODM)}

\item{ID}{index value for the specified object type}
}
\value{
A data frame
}
\description{
Set of utility functions to retrieve meta data from an ODM database
}
\details{
ODMgetCatalog()
Pass nothing - get a list of all datasets
Pass a SeriesID - get a single entry
}
\examples{

# Establish connection with database
ODM <- odbcConnect("ODM", "update", "update")

# Get a list of sites in the database
ODMgetSites()

# Get a list of data series in the database
ODMgetCatalog()

}
