% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMgetData.R
\name{ODMgetData}
\alias{ODMgetData}
\title{Query an ODM database}
\usage{
ODMgetData(SiteID_ = "SiteID", VariableID_ = "VariableID",
  MethodID_ = "MethodID", QualityControlLevelID_ = "QualityControlLevelID",
  AggregateBy = "day", FUN = "mean", startDate = NULL, endDate = NULL,
  channel = ODM)
}
\arguments{
\item{AggregateBy}{aggregation level to include in SQL query. Either hour,
day, month, year or none}

\item{FUN}{function to aggregate by. Either min, max, mean, or sum}

\item{startDate}{access data from this date forward}

\item{endDate}{access data up to this date}

\item{channel}{connection handle as returned by odbcConnect}

\item{SeriesID}{index value for a specific record within the series catalog}

\item{SiteID}{index value for the location at which the observation was made}

\item{VariableID}{index value for the variable that the data represents}

\item{MethodID}{index value for the method used to collect the observation}

\item{QualityCControlLevelID}{the level of quality control processing}
}
\value{
A data frame
}
\description{
The function ODMselect can be used to get data from ODM database. ODMselect
returns a data frame containing the required columns for working with ODM data.
}
\details{
A standard SQL query is issued to the ODM database and the values are
returned. The data returned can be limited by a start
and/or end date defined by the user. Smaller datasets can speed up the
workflow especially when visualising data.

Data series from the catalog may be specified using the series catalog
ID. Since the series catalog is not stable through time you may also query the
database by specifying SiteID, VariableID and MethodID.
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "update", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Extract data by site, variable, method and QC level. Clear and consistent
# method to extract data.
tmp <- ODMgetData(SiteID = 1, VariableID = 1, MethodID = 9,
 QualityCcontrolLevelID = 0, startDate = "2013-06-01", endDate = "2013-07-01",
 channel = ODM)
}

}
