% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMget.R
\name{ODMgetMethods}
\alias{ODMgetMethods}
\title{ODM utility functions}
\usage{
ODMgetMethods(ID = NA, Code = NA, channel = ODM)
}
\arguments{
\item{ID}{index value for the specified object type}

\item{Code}{code value for the specified object type}

\item{channel}{connection handle as returned by odbcConnect (default value = ODM)}
}
\value{
A data frame
}
\description{
Set of utility functions to retrieve meta data from an ODM database
}
\details{
ODMgetCatalog()
Pass nothing - get a list of all datasets
Pass a SeriesID - get a single entry
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "update", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Get a list of sites in the database
ODMgetSites()

# Get a list of data series in the database
ODMgetCatalog()
}

}
