% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMget.R
\name{ODMgetSites}
\alias{ODMgetSites}
\title{ODM utility functions}
\usage{
ODMgetSites(channel = ODM, ID = NA, Code = NA)
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect (default value = ODM)}

\item{ID}{index value for the specified object type}

\item{Code}{code value for the specified object type}
}
\value{
A data frame
}
\description{
Set of utility functions to retrieve meta data from an ODM database
}
\details{
ODMgetSites
Pass nothing - returns a full list of stations
Pass a MethodID - returns a single station
Pass a MethodCode - returns a single station
}
\examples{

# Establish connection with database
ODM <- odbcConnect("ODM", "update", "update")

# Get a list of sites in the database
ODMgetSites()

# Get a list of data series in the database
ODMgetCatalog()

}
