% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ODMgetSource}
\alias{ODMgetSource}
\title{ODM utility functions}
\usage{
ODMgetSource(channel = ODM, ID = NA)
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect (default value = ODM)}

\item{ID}{index value for the specified object type}
}
\value{
A data frame
}
\description{
Set of uility functions to retrieve meta data from an ODM database
}
\details{
ODMgetSource
Pass nothing - returns full list of sources
Pass a SourceID - returns a single source
}
\examples{

# Establish connection with database
ODM <- odbcConnect("ODM", "update", "update")

# Get a list of sites in the database
ODMgetSites()

# Get a list of data series in the database
ODMgetCatalog()

}
