% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMget.R
\name{ODMgetVariables}
\alias{ODMgetVariables}
\title{ODM utility functions}
\usage{
ODMgetVariables(channel = ODM, ID = NA, Code = NA)
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect (default value = ODM)}

\item{ID}{index value for the specified object type}

\item{Code}{code value for the specified object type}
}
\value{
A data frame
}
\description{
Set of utility functions to retrieve meta data from an ODM database
}
\details{
ODMgetVariables
Pass nothing - returns full list of variables
Pass a VariableID - returns a single variable
Pass a VariableCode - returns a single variable
}
\examples{

# Establish connection with database
ODM <- odbcConnect("ODM", "update", "update")

# Get a list of sites in the database
ODMgetSites()

# Get a list of data series in the database
ODMgetCatalog()

}
