% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query_load.R
\name{ODMload}
\alias{ODMload}
\title{Load data to an ODM database}
\usage{
ODMload(channel, obj, QCcheck = 1)
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect}

\item{obj}{xts object containing the appropriate data}

\item{QCcheck}{must match the quality control level of the data to be loaded}
}
\description{
The function ODMload can be used to get data into an ODM database. ODMload
takes an xts object containing the required columns for working with ODM data
and loads it. The xts objects will either have been acquired using ODMselect
or created using ODM create.
}
\details{
A standard SQL query is issued to the ODM database and the values are uploaded
in 100 row chunks. QClevel 1 data is handled by default but data containing
any QClevel data can be used by setting the argument to the appropriate value.
}
\examples{
\dontrun{ODM <- odbcConnect("Connection", "User id", "Password")
ODMload(ODM, "HP4_stage_corr")}
}

