% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMsummary.R
\name{ODMsummary}
\alias{ODMsummary}
\title{Summarise data formated for use with the ODM}
\usage{
ODMsummary(channel, Data)
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect}

\item{Data}{containing the data in a format consistent with the ODM data model}
}
\description{
The function ODMsummary can be used to verify the proper structure and use of
ID variables. Will return a meta data summary consistent with the
SeriesCatalog in the database.
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbcConnect("ODM", "update", "update")

# Return meta data for the data file
ODMsummary(ODM, HP4)
}

}
