% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMtools.R
\name{ODMtools}
\alias{ODMtools}
\title{ODMtools}
\usage{
ODMtools(...)
}
\description{
Interactively visualise data from an ODM database. R code can be executed and
data manipulated using the embedded Ace text editor.
}
