% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMupdateCatalog.R
\name{ODMupdateCatalog}
\alias{ODMupdateCatalog}
\title{Update ODM series catalog}
\usage{
ODMupdateCatalog(channel = ODM)
}
\description{
The function can be used to trigger the stored procedure in a MS-SQL
database for updating the series catalog.
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "sa", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Update catalog
ODMupdateCatalog()
}

}
