% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pint.R
\name{pint}
\alias{pint}
\title{Interactive plot (pint)}
\usage{
pint(x, y, data, geom = geom_point())
}
\arguments{
\item{x,y}{Name value pairs giving aesthetics to map}

\item{data}{Data frame to use}

\item{geom}{or list of geoms to draw. Defaults to geom_point()}
}
\description{
Visually explore data and highlight and interact with observations through
data brushing. A simple console allows for any R code to be run against
highlighted values including adding, removing, changing and summarizing.
}
\details{
Uses ggplot2 for producing the plots within a shiny app. The first two
arguments are "x" and "y", giving the x- and y-coordinates for the objects on
the plot. There is also a data argument used to declare the input data frame.
}
\examples{
df <- data.frame(x = rnorm(100), y = rnorm(100))
pint("x", "y", df)
pint("x","y",df, geom = geom_line())
}

