% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMload.R
\name{ODMload}
\alias{ODMload}
\title{Load data to an ODM database}
\usage{
ODMload(Data, QCcheck = 1, channel = ODM, batch_size = 1000,
  check_before_load = TRUE)
}
\arguments{
\item{Data}{containing the required columns}

\item{QCcheck}{must match the quality control level of the data to be loaded}

\item{channel}{connection handle as returned by odbcConnect}

\item{batch_size}{data is sent in small chunks to the database. This setting
decides the size of those chunks.}

\item{check_before_load}{will let you know what data you are about to load to
the database. You can turn this behavior off by setting this to FALSE.}
}
\description{
The function ODMload can be used to get data into the DataValues table in an
ODM database. ODMload takes a dataframe containing the required columns for
working with ODM data and loads it. The dataframes will either have been
acquired using ODMselect or created using ODMcreate.
}
\details{
A standard SQL query is issued to the ODM database and the values are uploaded
in 1000 row chunks. QClevel 1 data is handled by default but data containing
any QClevel data can be used by setting the argument to the appropriate value.
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "update", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Query the database
tmp <- ODMselect(SiteID = 1, VariableID = 1, MethodID = 9,
 QCLevelID = 0, startDate = "2013-06-01", endDate = "2013-07-01")

# Load values back to ODM
ODMload(Data = tmp, QCcheck = 0)
}

}
