% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMload.R
\name{ODMload}
\alias{ODMload}
\title{Load data to an ODM database}
\arguments{
\item{channel}{connection handle as returned by odbcConnect}

\item{Dataframe}{containing the required columns}

\item{QCcheck}{must match the quality control level of the data to be loaded}
}
\description{
The function ODMload can be used to get data into the DataValues table in an
ODM database. ODMload takes a dataframe containing the required columns for
working with ODM data and loads it. The dataframes will either have been
acquired using ODMselect or created using ODMcreate.
}
\details{
A standard SQL query is issued to the ODM database and the values are uploaded
in 100 row chunks. QClevel 1 data is handled by default but data containing
any QClevel data can be used by setting the argument to the appropriate value.
}
\examples{
# Establish connection with database
ODM <- odbcConnect("ODM", "update", "update")

# Query the database
tmp <- ODMselect(ODM, SiteID = 1, VariableID = 1, MethodID = 9,
 QCLevelID = 0, startDate = "2013-06-01", endDate = "2013-07-01")

# Load values back to ODM
ODMload(ODM, Data = tmp, QCcheck = 0)

}
