% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ODMselect.R
\name{ODMselect}
\alias{ODMselect}
\title{Query an ODM database}
\usage{
ODMselect(channel, CatalogID = NULL, SiteID = NULL, VariableID = NULL,
  MethodID = NULL, QCLevelID = 1, startDate = "1970-01-1 00:00:00",
  endDate = Sys.Date())
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect}

\item{CatalogID}{index value for a specific record within the series catalog}

\item{SiteID}{index value for the location at which the observation was made}

\item{VariableID}{index value for the variable that the data represents}

\item{MethodID}{index value for the method used to collect the observation}

\item{QCLevelID}{the level of quality control processing}

\item{startDate}{access data from this date forward}

\item{endDate}{access data up to this date}
}
\value{
An xts object
}
\description{
The function ODMselect can be used to get data from ODM database. ODMselect
returns an xts object containing required columns for working with ODM data.
The xts object contains an array of values comprising your data in matrix form
and an index based on the LocalDateTime.
}
\details{
A standard SQL query is issued to the ODM database and the values are
returned. Output is an xts object. The data returned can be limited by a start
and/or end date defined by the user. Smaller datasets can speed up the
workflow especially when visualising data.

A specific series from the catalog may be specified using the series catalog
ID. Since the series catalog is not stable through time you may also query the
database by specifying SiteID, VariableID and MethodID.
}
\examples{
\dontrun{ODM <- odbcConnect("Connection", "User id", "Password")
Data <- ODMSelect(ODM, CatalogID = 10, "2013-06-01", "2014-06-01")}
}

