% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMsummary.R
\name{ODMsummary}
\alias{ODMsummary}
\title{Summarise data formated for use with the ODM}
\usage{
ODMsummary(Data, channel = ODM)
}
\arguments{
\item{Data}{containing the data in a format consistent with the ODM data model}

\item{channel}{connection handle as returned by odbcConnect}
}
\description{
The function ODMsummary can be used to verify the proper structure and use of
ID variables. Will return a meta data summary consistent with the
SeriesCatalog in the database.
}
\examples{

\dontrun{
# Establish connection with database
ODM <- odbc::dbConnect(odbc::odbc(), dsn = "ODM", database = "OD",
 UID = "update", PWD = rstudioapi::askForPassword("Database password"),
 Port = 1433)

# Return meta data for the data file
ODMsummary(HP4, ODM)
}

}
